hwc: Disable WFD display when WFD and HDMI displays connected together
WFD and HDMI displays can't work together because they share a single DSS
manager. HDMI display has higher priority, so when HDMI display is
connected we should check if WFD display is running and disable it if
necessary. And we should disable WFD display after creation if HDMI
display is already connected.

Change-Id: I4a0b3a48ab9b7b05fa18c10b782d625d30bb3e2a
Signed-off-by: Maksym Kreshchuk <x0191602@ti.com>
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>