hwc: Fix display stack parsing logic
The mode of the display (whether presentation or legacy mode)
is identified by comparing the external display stack id
with that of the primary display. During the hotplug,
due to race condition, sometimes the stack id for the external display
is queried even before it is set from SurfaceFlinger.
This condition is handled by resetting the external display stack id
to the primary display stack id, if it is found to be invalid.

This fixes: Audio fails while hdmi plug/unplug
which was due to multiple setmodes happening from hwc during hdmi
hotplug. And the multiple setmodes were resulted from the invalid
stack id for the external display.

Change-Id: I205bd07e0388e23bad6fb76a449be145aacb8c79
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>